corechecks: Reorganize barrier related code
Move barrier related code from core_validation.cpp to
buffer_validation.cpp. Much of this code is going to be turned
into templates for synchronization2. It is much easier to manage
this if most of the template definitions and usages are in the
same compilation unit.
