+++ layers/core_validation.cpp
@@ -66,33 +66,6 @@
 #include "command_counter.h"
 #include "sync_utils.h"
 
-static VkImageLayout NormalizeImageLayout(VkImageLayout layout, VkImageLayout non_normal, VkImageLayout normal) {
-    return (layout == non_normal) ? normal : layout;
-}
-
-static VkImageLayout NormalizeDepthImageLayout(VkImageLayout layout) {
-    return NormalizeImageLayout(layout, VK_IMAGE_LAYOUT_DEPTH_STENCIL_READ_ONLY_OPTIMAL,
-                                VK_IMAGE_LAYOUT_DEPTH_READ_ONLY_STENCIL_ATTACHMENT_OPTIMAL);
-}
-
-static VkImageLayout NormalizeStencilImageLayout(VkImageLayout layout) {
-    return NormalizeImageLayout(layout, VK_IMAGE_LAYOUT_DEPTH_STENCIL_READ_ONLY_OPTIMAL,
-                                VK_IMAGE_LAYOUT_DEPTH_ATTACHMENT_STENCIL_READ_ONLY_OPTIMAL);
-}
-
-bool ImageLayoutMatches(const VkImageAspectFlags aspect_mask, VkImageLayout a, VkImageLayout b) {
-    bool matches = (a == b);
-    if (!matches) {
-        // Relaxed rules when referencing *only* the depth or stencil aspects
-        if (aspect_mask == VK_IMAGE_ASPECT_DEPTH_BIT) {
-            matches = NormalizeDepthImageLayout(a) == NormalizeDepthImageLayout(b);
-        } else if (aspect_mask == VK_IMAGE_ASPECT_STENCIL_BIT) {
-            matches = NormalizeStencilImageLayout(a) == NormalizeStencilImageLayout(b);
-        }
-    }
-    return matches;
-}
-
 // These functions are defined *outside* the core_validation namespace as their type
 // is also defined outside that namespace
 size_t PipelineLayoutCompatDef::hash() const {
@@ -7064,92 +7037,6 @@ static bool HasNonFramebufferStagePipelineStageFlags(VkPipelineStageFlags inflag
                         VK_PIPELINE_STAGE_LATE_FRAGMENT_TESTS_BIT | VK_PIPELINE_STAGE_COLOR_ATTACHMENT_OUTPUT_BIT)) != 0;
 }
 
-// Verify image barrier image state and that the image is consistent with FB image
-bool CoreChecks::ValidateImageBarrierAttachment(const char *funcName, CMD_BUFFER_STATE const *cb_state,
-                                                const FRAMEBUFFER_STATE *framebuffer, uint32_t active_subpass,
-                                                const safe_VkSubpassDescription2 &sub_desc, const VkRenderPass rp_handle,
-                                                uint32_t img_index, const VkImageMemoryBarrier &img_barrier,
-                                                const CMD_BUFFER_STATE *primary_cb_state) const {
-    bool skip = false;
-    const auto *fb_state = framebuffer;
-    assert(fb_state);
-    const auto img_bar_image = img_barrier.image;
-    bool image_match = false;
-    bool sub_image_found = false;  // Do we find a corresponding subpass description
-    VkImageLayout sub_image_layout = VK_IMAGE_LAYOUT_UNDEFINED;
-    uint32_t attach_index = 0;
-    // Verify that a framebuffer image matches barrier image
-    const auto attachment_count = fb_state->createInfo.attachmentCount;
-    for (uint32_t attachment = 0; attachment < attachment_count; ++attachment) {
-        auto view_state = GetActiveAttachmentImageViewState(cb_state, attachment, primary_cb_state);
-        if (view_state && (img_bar_image == view_state->create_info.image)) {
-            image_match = true;
-            attach_index = attachment;
-            break;
-        }
-    }
-    if (image_match) {  // Make sure subpass is referring to matching attachment
-        if (sub_desc.pDepthStencilAttachment && sub_desc.pDepthStencilAttachment->attachment == attach_index) {
-            sub_image_layout = sub_desc.pDepthStencilAttachment->layout;
-            sub_image_found = true;
-        }
-        if (!sub_image_found && device_extensions.vk_khr_depth_stencil_resolve) {
-            const auto *resolve = LvlFindInChain<VkSubpassDescriptionDepthStencilResolve>(sub_desc.pNext);
-            if (resolve && resolve->pDepthStencilResolveAttachment &&
-                resolve->pDepthStencilResolveAttachment->attachment == attach_index) {
-                sub_image_layout = resolve->pDepthStencilResolveAttachment->layout;
-                sub_image_found = true;
-            }
-        }
-        if (!sub_image_found) {
-            for (uint32_t j = 0; j < sub_desc.colorAttachmentCount; ++j) {
-                if (sub_desc.pColorAttachments && sub_desc.pColorAttachments[j].attachment == attach_index) {
-                    sub_image_layout = sub_desc.pColorAttachments[j].layout;
-                    sub_image_found = true;
-                    break;
-                }
-                if (!sub_image_found && sub_desc.pResolveAttachments &&
-                    sub_desc.pResolveAttachments[j].attachment == attach_index) {
-                    sub_image_layout = sub_desc.pResolveAttachments[j].layout;
-                    sub_image_found = true;
-                    break;
-                }
-            }
-        }
-        if (!sub_image_found) {
-            skip |= LogError(rp_handle, "VUID-vkCmdPipelineBarrier-image-04073",
-                             "%s: Barrier pImageMemoryBarriers[%d].%s is not referenced by the VkSubpassDescription for "
-                             "active subpass (%d) of current %s.",
-                             funcName, img_index, report_data->FormatHandle(img_bar_image).c_str(), active_subpass,
-                             report_data->FormatHandle(rp_handle).c_str());
-        }
-    } else {  // !image_match
-        skip |=
-            LogError(fb_state->framebuffer, "VUID-vkCmdPipelineBarrier-image-04073",
-                     "%s: Barrier pImageMemoryBarriers[%d].%s does not match an image from the current %s.", funcName, img_index,
-                     report_data->FormatHandle(img_bar_image).c_str(), report_data->FormatHandle(fb_state->framebuffer).c_str());
-    }
-    if (img_barrier.oldLayout != img_barrier.newLayout) {
-        skip |= LogError(cb_state->commandBuffer, "VUID-vkCmdPipelineBarrier-oldLayout-01181",
-                         "%s: As the Image Barrier for %s is being executed within a render pass instance, oldLayout must "
-                         "equal newLayout yet they are %s and %s.",
-                         funcName, report_data->FormatHandle(img_barrier.image).c_str(),
-                         string_VkImageLayout(img_barrier.oldLayout), string_VkImageLayout(img_barrier.newLayout));
-    } else {
-        if (sub_image_found && sub_image_layout != img_barrier.oldLayout) {
-            LogObjectList objlist(rp_handle);
-            objlist.add(img_bar_image);
-            skip |= LogError(objlist, "VUID-vkCmdPipelineBarrier-oldLayout-01181",
-                             "%s: Barrier pImageMemoryBarriers[%d].%s is referenced by the VkSubpassDescription for active "
-                             "subpass (%d) of current %s as having layout %s, but image barrier has layout %s.",
-                             funcName, img_index, report_data->FormatHandle(img_bar_image).c_str(), active_subpass,
-                             report_data->FormatHandle(rp_handle).c_str(), string_VkImageLayout(sub_image_layout),
-                             string_VkImageLayout(img_barrier.oldLayout));
-        }
-    }
-    return skip;
-}
-
 // Validate image barriers within a renderPass
 bool CoreChecks::ValidateRenderPassImageBarriers(const char *funcName, const CMD_BUFFER_STATE *cb_state, uint32_t active_subpass,
                                                  const safe_VkSubpassDescription2 &sub_desc, const VkRenderPass rp_handle,
@@ -7322,227 +7209,6 @@ static bool ValidateAccessMaskPipelineStage(const DeviceFeatures &features, VkQu
     auto expanded_stages = sync_utils::ExpandPipelineStages(stage_mask, queue_flags, sync_utils::DisabledPipelineStages(features));
     auto valid_accesses = sync_utils::CompatibleAccessMask(expanded_stages);
     return ((access_mask & ~valid_accesses) == 0);
-}
-
-namespace barrier_queue_families {
-enum VuIndex {
-    kSrcOrDstMustBeIgnore,
-    kSpecialOrIgnoreOnly,
-    kSrcAndDstValidOrSpecial,
-    kSrcAndDestMustBeIgnore,
-    kSrcAndDstBothValid,
-    kSubmitQueueMustMatchSrcOrDst
-};
-static const char *vu_summary[] = {"Source or destination queue family must be ignored.",
-                                   "Source or destination queue family must be special or ignored.",
-                                   "Destination queue family must be ignored if source queue family is.",
-                                   "Destination queue family must be valid, ignored, or special.",
-                                   "Source queue family must be valid, ignored, or special.",
-                                   "Source and destination queue family must both be ignored.",
-                                   "Source and destination queue family must both be ignore or both valid.",
-                                   "Source or destination queue family must match submit queue family, if not ignored."};
-
-static const std::string kImageErrorCodes[] = {
-    "VUID-VkImageMemoryBarrier-synchronization2-03857",                                   //   kSrcOrDstMustBeIgnore
-    "VUID-VkImageMemoryBarrier-image-04071",                                              //   kSpecialOrIgnoreOnly
-    "VUID-VkImageMemoryBarrier-image-04072",                                              //   kSrcAndDstValidOrSpecial
-    "VUID-VkImageMemoryBarrier-synchronization2-03856",                                   //   kSrcAndDestMustBeIgnore
-    "VUID-VkImageMemoryBarrier-image-04069",                                              //   kSrcAndDstBothValid
-    "UNASSIGNED-CoreValidation-vkImageMemoryBarrier-sharing-mode-exclusive-same-family",  //   kSubmitQueueMustMatchSrcOrDst
-};
-
-static const std::string kBufferErrorCodes[] = {
-    "VUID-VkBufferMemoryBarrier-synchronization2-03853",                                   //  kSrcOrDstMustBeIgnore
-    "VUID-VkBufferMemoryBarrier-buffer-04088",                                             //  kSpecialOrIgnoreOnly
-    "VUID-VkBufferMemoryBarrier-buffer-04089",                                             //  kSrcAndDstValidOrSpecial
-    "VUID-VkBufferMemoryBarrier-synchronization2-03852",                                   //  kSrcAndDestMustBeIgnore
-    "VUID-VkBufferMemoryBarrier-buffer-04086",                                             //  kSrcAndDstBothValid
-    "UNASSIGNED-CoreValidation-vkBufferMemoryBarrier-sharing-mode-exclusive-same-family",  //  kSubmitQueueMustMatchSrcOrDst
-};
-
-class ValidatorState {
-  public:
-    ValidatorState(const ValidationStateTracker *device_data, const char *func_name, const CMD_BUFFER_STATE *cb_state,
-                   const VulkanTypedHandle &barrier_handle, const VkSharingMode sharing_mode)
-        : device_data_(device_data),
-          func_name_(func_name),
-          command_buffer_(cb_state->commandBuffer),
-          barrier_handle_(barrier_handle),
-          sharing_mode_(sharing_mode),
-          val_codes_(barrier_handle.type == kVulkanObjectTypeImage ? kImageErrorCodes : kBufferErrorCodes),
-          limit_(static_cast<uint32_t>(device_data->physical_device_state->queue_family_properties.size())),
-          mem_ext_(IsExtEnabled(device_data->device_extensions.vk_khr_external_memory)) {}
-
-    // Log the messages using boilerplate from object state, and Vu specific information from the template arg
-    // One and two family versions, in the single family version, Vu holds the name of the passed parameter
-    bool LogMsg(VuIndex vu_index, uint32_t family, const char *param_name) const {
-        const std::string &val_code = val_codes_[vu_index];
-        const char *annotation = GetFamilyAnnotation(family);
-        return device_data_->LogError(command_buffer_, val_code,
-                                      "%s: Barrier using %s %s created with sharingMode %s, has %s %u%s. %s", func_name_,
-                                      GetTypeString(), device_data_->report_data->FormatHandle(barrier_handle_).c_str(),
-                                      GetModeString(), param_name, family, annotation, vu_summary[vu_index]);
-    }
-
-    bool LogMsg(VuIndex vu_index, uint32_t src_family, uint32_t dst_family) const {
-        const std::string &val_code = val_codes_[vu_index];
-        const char *src_annotation = GetFamilyAnnotation(src_family);
-        const char *dst_annotation = GetFamilyAnnotation(dst_family);
-        return device_data_->LogError(
-            command_buffer_, val_code,
-            "%s: Barrier using %s %s created with sharingMode %s, has srcQueueFamilyIndex %u%s and dstQueueFamilyIndex %u%s. %s",
-            func_name_, GetTypeString(), device_data_->report_data->FormatHandle(barrier_handle_).c_str(), GetModeString(),
-            src_family, src_annotation, dst_family, dst_annotation, vu_summary[vu_index]);
-    }
-
-    // This abstract Vu can only be tested at submit time, thus we need a callback from the closure containing the needed
-    // data. Note that the mem_barrier is copied to the closure as the lambda lifespan exceed the guarantees of validity for
-    // application input.
-    static bool ValidateAtQueueSubmit(const QUEUE_STATE *queue_state, const ValidationStateTracker *device_data,
-                                      uint32_t src_family, uint32_t dst_family, const ValidatorState &val) {
-        uint32_t queue_family = queue_state->queueFamilyIndex;
-        if ((src_family != queue_family) && (dst_family != queue_family)) {
-            const std::string &val_code = val.val_codes_[kSubmitQueueMustMatchSrcOrDst];
-            const char *src_annotation = val.GetFamilyAnnotation(src_family);
-            const char *dst_annotation = val.GetFamilyAnnotation(dst_family);
-            return device_data->LogError(
-                queue_state->queue, val_code,
-                "%s: Barrier submitted to queue with family index %u, using %s %s created with sharingMode %s, has "
-                "srcQueueFamilyIndex %u%s and dstQueueFamilyIndex %u%s. %s",
-                "vkQueueSubmit", queue_family, val.GetTypeString(),
-                device_data->report_data->FormatHandle(val.barrier_handle_).c_str(), val.GetModeString(), src_family,
-                src_annotation, dst_family, dst_annotation, vu_summary[kSubmitQueueMustMatchSrcOrDst]);
-        }
-        return false;
-    }
-    // Logical helpers for semantic clarity
-    inline bool KhrExternalMem() const { return mem_ext_; }
-    inline bool IsValid(uint32_t queue_family) const { return (queue_family < limit_); }
-    inline bool IsValidOrSpecial(uint32_t queue_family) const {
-        return IsValid(queue_family) || (mem_ext_ && QueueFamilyIsExternal(queue_family));
-    }
-
-    // Helpers for LogMsg
-    const char *GetModeString() const { return string_VkSharingMode(sharing_mode_); }
-
-    // Descriptive text for the various types of queue family index
-    const char *GetFamilyAnnotation(uint32_t family) const {
-        const char *external = " (VK_QUEUE_FAMILY_EXTERNAL)";
-        const char *foreign = " (VK_QUEUE_FAMILY_FOREIGN_EXT)";
-        const char *ignored = " (VK_QUEUE_FAMILY_IGNORED)";
-        const char *valid = " (VALID)";
-        const char *invalid = " (INVALID)";
-        switch (family) {
-            case VK_QUEUE_FAMILY_EXTERNAL:
-                return external;
-            case VK_QUEUE_FAMILY_FOREIGN_EXT:
-                return foreign;
-            case VK_QUEUE_FAMILY_IGNORED:
-                return ignored;
-            default:
-                if (IsValid(family)) {
-                    return valid;
-                }
-                return invalid;
-        };
-    }
-    const char *GetTypeString() const { return object_string[barrier_handle_.type]; }
-    VkSharingMode GetSharingMode() const { return sharing_mode_; }
-
-  protected:
-    const ValidationStateTracker *device_data_;
-    const char *const func_name_;
-    const VkCommandBuffer command_buffer_;
-    const VulkanTypedHandle barrier_handle_;
-    const VkSharingMode sharing_mode_;
-    const std::string *val_codes_;
-    const uint32_t limit_;
-    const bool mem_ext_;
-};
-
-bool Validate(const CoreChecks *device_data, const char *func_name, const CMD_BUFFER_STATE *cb_state, const ValidatorState &val,
-              const uint32_t src_queue_family, const uint32_t dst_queue_family) {
-    bool skip = false;
-
-    const bool mode_concurrent = val.GetSharingMode() == VK_SHARING_MODE_CONCURRENT;
-    const bool src_ignored = QueueFamilyIsIgnored(src_queue_family);
-    const bool dst_ignored = QueueFamilyIsIgnored(dst_queue_family);
-    if (val.KhrExternalMem()) {
-        if (mode_concurrent) {
-            if (!(src_ignored || dst_ignored)) {
-                skip |= val.LogMsg(kSrcOrDstMustBeIgnore, src_queue_family, dst_queue_family);
-            }
-            if ((src_ignored && !(dst_ignored || QueueFamilyIsExternal(dst_queue_family))) ||
-                (dst_ignored && !(src_ignored || QueueFamilyIsExternal(src_queue_family)))) {
-                skip |= val.LogMsg(kSpecialOrIgnoreOnly, src_queue_family, dst_queue_family);
-            }
-        } else {
-            // VK_SHARING_MODE_EXCLUSIVE
-            if (src_queue_family != dst_queue_family) {
-                if (!val.IsValidOrSpecial(dst_queue_family)) {
-                    skip |= val.LogMsg(kSrcAndDstValidOrSpecial, dst_queue_family, "dstQueueFamilyIndex");
-                }
-                if (!val.IsValidOrSpecial(src_queue_family)) {
-                    skip |= val.LogMsg(kSrcAndDstValidOrSpecial, src_queue_family, "srcQueueFamilyIndex");
-                }
-            }
-        }
-    } else {
-        // No memory extension
-        if (mode_concurrent) {
-            if (!src_ignored || !dst_ignored) {
-                skip |= val.LogMsg(kSrcAndDestMustBeIgnore, src_queue_family, dst_queue_family);
-            }
-        } else {
-            // VK_SHARING_MODE_EXCLUSIVE
-            if ((src_queue_family != dst_queue_family) && !(val.IsValid(src_queue_family) && val.IsValid(dst_queue_family))) {
-                skip |= val.LogMsg(kSrcAndDstBothValid, src_queue_family, dst_queue_family);
-            }
-        }
-    }
-    return skip;
-}
-}  // namespace barrier_queue_families
-
-bool CoreChecks::ValidateConcurrentBarrierAtSubmit(const ValidationStateTracker *state_data, const QUEUE_STATE *queue_state,
-                                                   const char *func_name, const CMD_BUFFER_STATE *cb_state,
-                                                   const VulkanTypedHandle &typed_handle, uint32_t src_queue_family,
-                                                   uint32_t dst_queue_family) {
-    using barrier_queue_families::ValidatorState;
-    ValidatorState val(state_data, func_name, cb_state, typed_handle, VK_SHARING_MODE_CONCURRENT);
-    return ValidatorState::ValidateAtQueueSubmit(queue_state, state_data, src_queue_family, dst_queue_family, val);
-}
-
-// Type specific wrapper for image barriers
-bool CoreChecks::ValidateBarrierQueueFamilies(const char *func_name, const CMD_BUFFER_STATE *cb_state,
-                                              const VkImageMemoryBarrier &barrier, const IMAGE_STATE *state_data) const {
-    // State data is required
-    if (!state_data) {
-        return false;
-    }
-
-    // Create the validator state from the image state
-    barrier_queue_families::ValidatorState val(this, func_name, cb_state, VulkanTypedHandle(barrier.image, kVulkanObjectTypeImage),
-                                               state_data->createInfo.sharingMode);
-    const uint32_t src_queue_family = barrier.srcQueueFamilyIndex;
-    const uint32_t dst_queue_family = barrier.dstQueueFamilyIndex;
-    return barrier_queue_families::Validate(this, func_name, cb_state, val, src_queue_family, dst_queue_family);
-}
-
-// Type specific wrapper for buffer barriers
-bool CoreChecks::ValidateBarrierQueueFamilies(const char *func_name, const CMD_BUFFER_STATE *cb_state,
-                                              const VkBufferMemoryBarrier &barrier, const BUFFER_STATE *state_data) const {
-    // State data is required
-    if (!state_data) {
-        return false;
-    }
-
-    // Create the validator state from the buffer state
-    barrier_queue_families::ValidatorState val(
-        this, func_name, cb_state, VulkanTypedHandle(barrier.buffer, kVulkanObjectTypeBuffer), state_data->createInfo.sharingMode);
-    const uint32_t src_queue_family = barrier.srcQueueFamilyIndex;
-    const uint32_t dst_queue_family = barrier.dstQueueFamilyIndex;
-    return barrier_queue_families::Validate(this, func_name, cb_state, val, src_queue_family, dst_queue_family);
 }
 
 bool CoreChecks::ValidateBarriers(const char *funcName, const CMD_BUFFER_STATE *cb_state, VkPipelineStageFlags src_stage_mask,
